update: indicators.py, signals_plot.py "created class for generate all indicators and plots."
