+++ indicator/indicators.py
@@ -4,6 +4,8 @@ from moving_average_signal import MovingAverage
 from bollinger_bands import BollingerBand
 from macd import Macd
 from rsi import Rsi
+from super_trend import SuperTrend
+from candle_pattern import MakePattern
 
 class CreateIndicators:
     def __init__(self, data):
@@ -34,9 +36,49 @@ class CreateIndicators:
 
     def super_trend(self):
         # It will return the signal super_trend
-        pass
+        st = SuperTrend()
+        df = st.create_super_trend(self.data)
+        return df['signal']
 
     def candle_pattern(self):
         # It will return the signal candle_pattern
-        pass
+        mp = MakePattern()
+        pattern = mp.pattern(self.data)
+        return pattern
+    def create_all_indicators(self):
+        bd = self.bollinger_band()
+        ma = self.moving_average()
+        macd = self.macd()
+        rsi = self.rsi()
+        st = self.super_trend()
+        cp = self.candle_pattern()
 
+        df = pd.DataFrame({'bollinger_band': bd,
+                           'moving_average': ma,
+                           'macd': macd,
+                           'rsi': rsi,
+                           'super_trend': st})
+        df = df.add(cp, fill_value=0)
+        df = df.astype(int)
+        return df
+
+if __name__ == "__main__":
+    from database.dataframe import GetDataframe
+    import matplotlib.pyplot as plt
+    data = GetDataframe().get_minute_data('BTCBUSD', 1, 1000)
+    ci = CreateIndicators(data)
+    print("All Indicators: ")
+    df = ci.create_all_indicators()
+    print(df)
+    data['sum'] = df.sum(axis=1)
+
+    print(data)
+
+    # Making Plot for batter visualization
+    plt.plot(data['Close'], label='Close Price')
+    plt.scatter(data.index[data['sum'] >= 100], data['Close'][data['sum'] >= 100],
+                marker='^', s=20, color='green', label='Buy signal', zorder=3)
+    plt.scatter(data.index[data['sum'] <= -100], data['Close'][data['sum'] <= -100],
+                marker='v', s=20, color='red', label='Sell signal', zorder=3)
+    plt.legend()
+    plt.show()
